# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for DecentraRent - Solana dApp for asset renting",
        "title": "DecentraRent API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all conversations for the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "List conversations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chat.Conversation"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new conversation between the authenticated user (loaner) and a landlord for a property, or returns the existing one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Create or get a conversation",
                "parameters": [
                    {
                        "description": "Conversation details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.createConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chat.Conversation"
                        }
                    }
                }
            }
        },
        "/conversations/documents": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sends a document (agreement PDF, etc.) in a conversation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Send a document in chat",
                "parameters": [
                    {
                        "description": "Document details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/chat.sendDocumentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/chat.Message"
                        }
                    }
                }
            }
        },
...
        "/conversations/{id}/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns messages for a conversation with pagination. Supports all message types: text, system, document, modal.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "Get messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Load messages before this timestamp (RFC3339)",
                        "name": "before",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/chat.Message"
                            }
                        }
                    }
                }
            }
        },
...
        "/properties/{id}/media": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers a media file after it has been uploaded to MinIO (owner only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Register uploaded media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "File key from upload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/media.RegisterMediaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/media.PropertyMedia"
                        }
                    }
                }
            }
        },
        "/properties/{id}/media/order": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the sort order of media items (owner only)",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Reorder media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ordered list of media IDs",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/media.ReorderRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/properties/{id}/media/upload-url": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a presigned MinIO URL for direct file upload (owner only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "Get presigned upload URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "File name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/media.UploadURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/media.UploadURLResponse"
                        }
                    }
                }
            }
        },
        "/properties/{id}/media/{mediaId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a media item and its S3 object (owner only)",
                "tags": [
                    "media"
                ],
                "summary": "Delete media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Property ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Media ID",
                        "name": "mediaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "chat.Conversation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "landlord_id": {
                    "type": "string"
                },
                "last_message": {
                    "type": "string"
                },
                "last_message_at": {
                    "type": "string"
                },
                "loaner_id": {
                    "type": "string"
                },
                "property_id": {
                    "type": "string"
                }
            }
        },
        "chat.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message_type": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/chat.MessageMeta"
                },
                "sender_id": {
                    "type": "string"
                }
            }
        },
        "chat.MessageMeta": {
            "type": "object",
            "properties": {
                "document_name": {
                    "type": "string"
                },
                "document_url": {
                    "description": "Для document",
                    "type": "string"
                },
                "event_type": {
                    "description": "Для system",
                    "type": "string"
                },
                "modal_action": {
                    "description": "Для modal",
                    "type": "string"
                },
                "modal_id": {
                    "description": "unique ID для отслеживания ответа",
                    "type": "string"
                },
                "modal_status": {
                    "description": "pending, accepted, rejected",
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                }
            }
        },
        "chat.createConversationRequest": {
            "type": "object",
            "properties": {
                "landlord_id": {
                    "type": "string"
                },
                "property_id": {
                    "type": "string"
                }
            }
        },
        "chat.sendDocumentRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "описание",
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "document_name": {
                    "description": "имя файла",
                    "type": "string"
                },
                "document_url": {
                    "description": "ссылка на файл",
                    "type": "string"
                },
                "order_id": {
                    "description": "привязка к ордеру",
                    "type": "string"
                }
            }
        },
...
        "media.PropertyMedia": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_key": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "property_id": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "media.RegisterMediaRequest": {
            "type": "object",
            "properties": {
                "file_key": {
                    "type": "string"
                }
            }
        },
        "media.ReorderRequest": {
            "type": "object",
            "properties": {
                "order": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "media.UploadURLRequest": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                }
            }
        },
        "media.UploadURLResponse": {
            "type": "object",
            "properties": {
                "file_key": {
                    "type": "string"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}